tp3/requirements.txt
boto3
python-dotenv
paramiko
matplotlib
numpy
scp